+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 - Added Layer Switcher to map (now with HeiGIT Carto and Bing Aerial maps), with opacity slider for GeoTIFF/GeoJSON layers from artifacts
 - Search function to help with finding PoIs during compute, and locations within results ([#179](https://gitlab.heigit.org/climate-action/web-app/-/issues/179))
 
+- Utilize `geojson-vt` library to render GeoJSON layers in a more performant way
+- Added visual regression testing with Cypress
+
 ## [1.3.0](https://gitlab.heigit.org/climate-action/web-app/-/releases/1.3.0) - 2024-10-14
 
 ### Changed